+++ packages/backend/package.json
@@ -18,6 +18,7 @@
   },
   "dependencies": {
     "@backstage/backend-common": "^0.19.8",
+    "@backstage/backend-defaults": "^0.2.6",
     "@backstage/backend-tasks": "^0.5.11",
     "@backstage/catalog-client": "^1.4.5",
     "@backstage/catalog-model": "^1.4.3",
